Changes for running plugin in Mura 6.2
Ran into several issues while attempting to run this plugin on Mura 6.2. Changes for this file:

Added a trailing slash to the "this.muraAppConfigPath" variable in the pre 7.1 section.

Added setting of the request timeout to extend it. This process takes a long time on a server with several pages.  I set it to 60 seconds here but it was taking over 1.5 hours on our production server. It will need to be set per installation or configurable perhaps.

Added setting of the "session.siteid" because it was being set to "default" when running from the scheduled task.  I assume the session does not exist when running from the scheduler.  The condition checks for the default setting, and if true then check to see if the "url.site" has been set and uses that.  Otherwise stick with "default".