Speed up the indexing process
On our site of roughly 5,000 pages this plugin was taking over 1.5 hours to run.  With the suggested edit the process now takes around 30 minutes.  Basically the change is to only loop over approved and active pages.  In our environment, before the change the qAtts query was returning 43,165 rows.  After the change it only returns 12,469 rows.  The resulting sitemap.xml file is identical in our environment.  Please verify and confirm in your environment.