Changes for running plugin in Mura 6.2
The Mura script function does not exist in Mura 6.2. So the CSS needs to be loaded another way. I used the same check logic that is being used in the Application.cfc file here.
Also, the script statement was outside of the HTML body so I moved it up within the savecontent block.